Recreate tree fix
When syncing a single item - recreate tree param is set to run
same behaviour as before

When syncing multiple items by type - recreate tree param is set
to false, instead the recreate tree button should be pressed after.

Same for sync all.

In draft, it is the upsert that is executed.
In live, it is the copy that is executed.

Both these suppport a parameter if recreate tree should be run or no.