Added checkbox to control indexing (0.14.20-dupont).
This is a new feature and to preserve backwards compatibility
a new parameter has been added to domain-settings entry.

It's called doNotIndex, and is thus undefined in existing databases.

In the UI, there is now a checkbox that can be checked (and unchecked).

If the checkbox is checked, we set doNotIndex to true.